Desktop_Vision_Assistant/requirements.txt
# Core dependencies
numpy>=1.22.0
opencv-python>=4.7.0
PyQt6>=6.4.0
pillow>=9.4.0
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.65.0
colorlog>=6.7.0

# Speech recognition
openai>=1.79.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Computer vision
ultralytics>=8.0.0  # YOLOv8

# Basic NLP (simplified to avoid build issues)
transformers>=4.30.0